refactor: renmae host module to ports
